Update deprecated php type casts
In PHP 8.5, non-canonical scalar type casts (boolean|double|integer|binary) are deprecated.

See https://php.watch/versions/8.5/boolean-double-integer-binary-casts-deprecated

+++ CRM/Banking/Rules/Rule.php
@@ -637,7 +637,7 @@ class CRM_Banking_Rules_Rule {
       case 'amount_min':
       case 'amount_max':
         // Floats
-        $this->$prop = (double) $value;
+        $this->$prop = (float) $value;
         break;
 
       default:
